Move qname classes into another package
