+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/ast/JavaQualifiableNode.java
@@ -6,6 +6,7 @@ package net.sourceforge.pmd.lang.java.ast;
 
 import net.sourceforge.pmd.lang.ast.QualifiableNode;
 
+
 /**
  * Java nodes that can be described with a qualified name.
  *
