+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/multifile/ProjectMirror.java
@@ -4,8 +4,8 @@
 
 package net.sourceforge.pmd.lang.java.multifile;
 
-import net.sourceforge.pmd.lang.java.ast.JavaOperationQualifiedName;
-import net.sourceforge.pmd.lang.java.ast.JavaTypeQualifiedName;
+import net.sourceforge.pmd.lang.java.qname.JavaOperationQualifiedName;
+import net.sourceforge.pmd.lang.java.qname.JavaTypeQualifiedName;
 import net.sourceforge.pmd.lang.java.multifile.signature.JavaFieldSigMask;
 import net.sourceforge.pmd.lang.java.multifile.signature.JavaOperationSigMask;
