+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/multifile/PackageStatsTest.java
@@ -16,9 +16,9 @@ import org.junit.Test;
 import net.sourceforge.pmd.lang.java.ast.ASTFieldDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodDeclaration;
 import net.sourceforge.pmd.lang.java.ast.ASTMethodOrConstructorDeclaration;
-import net.sourceforge.pmd.lang.java.ast.JavaOperationQualifiedName;
-import net.sourceforge.pmd.lang.java.ast.JavaTypeQualifiedName;
-import net.sourceforge.pmd.lang.java.ast.QualifiedNameFactory;
+import net.sourceforge.pmd.lang.java.qname.JavaOperationQualifiedName;
+import net.sourceforge.pmd.lang.java.qname.JavaTypeQualifiedName;
+import net.sourceforge.pmd.lang.java.qname.QualifiedNameFactory;
 import net.sourceforge.pmd.lang.java.multifile.signature.JavaFieldSigMask;
 import net.sourceforge.pmd.lang.java.multifile.signature.JavaFieldSignature;
 import net.sourceforge.pmd.lang.java.multifile.signature.JavaOperationSigMask;
